Final Commit
 So for my theme I decided to do an obstacle course! I added three different features into it.

First I added a timer to the game- so you have to collect all of the pickups in about 20 seconds in order to win. If you win, the game would pause the timer. If you lose the game closes. 

Second I added and composed a music track. It took me an hour or so to get the summer sound I wanted- a sort of vibey nostalgic thing and put that into the game.

Third, I added a sort of anti-pickup that makes you lose time if you come into contact with it. This is sort of the "obstacle" in the obstacle course theme!